Prep diff: code reorganization
