+++ docs/whats-new.md
@@ -46,7 +46,7 @@ Date: Apr 24, 2018
 
 ## Transform class (New, WebGL2) (Experimental)
 
-The new experimental [`Transform`](/docs/api-reference/core/experimental/transform.md) class provides an easy-to-use interface to perform Transform Feedback operations.
+The new experimental [`Transform`](/docs/api-reference/core/transform.md) class provides an easy-to-use interface to perform Transform Feedback operations.
 
 
 ## Framebuffer Class
