+++ src/core/shader-cache.js
@@ -1,8 +1,6 @@
+import {VertexShader, FragmentShader} from '../webgl/shader';
+import Program from '../webgl/program';
 import assert from '../utils/assert';
-
-// TODO - luma dependencies, refactor to make shadertools independent
-import {VertexShader, FragmentShader} from '../../../webgl/shader';
-import Program from '../../../webgl/program';
 
 export default class ShaderCache {
