+++ test/src/core/index.js
@@ -1,2 +1,2 @@
 // Experimental
-import './experimental/transform.spec';
+import './transform.spec';
